Moved reducers and model tests
Moved the following tests from the tests folder to their own
subdirectory: online_learning, regression_learning, model_filter,
feature_selection, component_analyzers, and classification_learning.
