+++ modules/DMZ/models/test/test_classification_learning.py
@@ -4,7 +4,7 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath('..'))
+sys.path.insert(0, os.path.abspath('../../../..'))
 
 import unittest
 from modules.DMZ.models.classification import scikit_classification_learners
