+++ modules/DMZ/models/test/test_model_filter.py
@@ -1,7 +1,7 @@
 from __future__ import division # Forcing floating point division
 import sys
 import os
-sys.path.insert(0, os.path.abspath('..'))
+sys.path.insert(0, os.path.abspath('../../../..'))
 
 from unittest import TestCase
 from modules.DMZ.models import model_filter
