+++ modules/DMZ/models/test/test_regression_learning.py
@@ -1,34 +1,35 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath('..'))
+sys.path.insert(0, os.path.abspath('../../../..'))
 
 from unittest import TestCase
-from modules.DMZ.models.regression import scikit_online_regressors
+from modules.DMZ.models.regression import scikit_regression_learners
 from modules.toolbox import ml_runners as mr
+
 from modules.toolbox.data_package import DataPackage
 from modules.DMZ.data_kit.validation_package import ValidationPackage
 
 
-class TestOnlineLearning(TestCase):
+class TestRegressionLearning(TestCase):
 
-    def test_sgd_regressor(self):
+    def test_bayesian_ridge(self):
         # Arrange
         data, validation_pack = self.setup_data()
 
         # Act
-        model = scikit_online_regressors.train_sgd_regressor()
+        model = scikit_regression_learners.train_bayesian_ridge()
         result = mr.model_use(model[1], validation_pack, data)
 
         # Assert
         self.assertGreater(result, 0, msg="Failed to beat baseline")
-    
-    def test_passive_aggressive_regressor(self):
+
+    def test_support_vector_regression(self):
         # Arrange
         data, validation_pack = self.setup_data()
 
         # Act
-        model = scikit_online_regressors.train_passive_aggressive_regressor()
+        model = scikit_regression_learners.train_support_vector_regression()
         result = mr.model_use(model[1], validation_pack, data)
 
         # Assert
