+++ modules/reducers/test/test_feature_selection.py
@@ -1,7 +1,7 @@
 import os
 import sys
 
-sys.path.insert(0, os.path.abspath('..'))
+sys.path.insert(0, os.path.abspath('../../..'))
 
 from unittest import TestCase
 from modules.DMZ.data_kit import data_io
